backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
redis==5.0.8
rq==1.16.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.9
PyPDF2==3.0.1
python-dotenv==1.0.1
openai==0.28.1


